rename: model -> domain 폴더명 변경
